+++ src/actions/sources/prettyPrint.js
@@ -21,7 +21,7 @@ import {
   getSelectedLocation
 } from "../../selectors";
 
-import type { ThunkArgs } from "../types";
+import type { Action, ThunkArgs } from "../types";
 
 export function createPrettySource(sourceId: string) {
   return async ({ dispatch, getState, sourceMaps }: ThunkArgs) => {
@@ -36,7 +36,7 @@ export function createPrettySource(sourceId: string) {
       contentType: "text/javascript",
       loadedState: "loading"
     };
-    dispatch({ type: "ADD_SOURCE", source: prettySource });
+    dispatch(({ type: "ADD_SOURCE", source: prettySource }: Action));
 
     const { code, mappings } = await prettyPrint({ source, url });
     await sourceMaps.applySourceMap(source.id, url, code, mappings);
@@ -49,7 +49,7 @@ export function createPrettySource(sourceId: string) {
 
     setSource(loadedPrettySource);
 
-    dispatch({ type: "UPDATE_SOURCE", source: loadedPrettySource });
+    dispatch(({ type: "UPDATE_SOURCE", source: loadedPrettySource }: Action));
 
     return prettySource;
   };
