+++ src/reducers/tests/sources.spec.js
@@ -5,6 +5,8 @@ declare var expect: (value: any) => any;
 
 import update, { initialSourcesState } from "../sources";
 import { foobar } from "../../test/fixtures";
+import type { Source } from "../../types";
+
 const fakeSources = foobar.sources.sources;
 
 describe("sources reducer", () => {
@@ -12,7 +14,8 @@ describe("sources reducer", () => {
     let state = initialSourcesState();
     state = update(state, {
       type: "ADD_SOURCE",
-      source: fakeSources.fooSourceActor
+      // coercing to a Source for the purpose of this test
+      source: ((fakeSources.fooSourceActor: any): Source)
     });
     expect(state.sources.size).toBe(1);
   });
